refactor: code gen has its own sub module now
