+++ src/code_gen/mod.rs
@@ -1,235 +1,18 @@
-use crate::config::{Count, Method, Service, Severity, Task};
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum StackValue {
-    String(String),
-    Int(u64),
-}
-
-impl std::fmt::Display for StackValue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            StackValue::String(s) => write!(f, "{}", s),
-            StackValue::Int(n) => write!(f, "{}", n),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Instruction {
-    Push(StackValue),
-    Pop,
-    Dec,
-    JmpIfZero(String),
-    Label(String),
-    Stdout,
-    Stderr,
-    Sleep(u64),
-    StoreVar(String, String),
-    LoadVar(String),
-    Dup,
-    Jump(String),
-    Printf,
-    RemoteCall,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ByteCodeError {
-    UnsupportedConst(String),
-}
-
-impl std::error::Error for ByteCodeError {}
-
-impl std::fmt::Display for ByteCodeError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ByteCodeError::UnsupportedConst(val) => write!(f, "Unsupported constant: {}", val),
-        }
-    }
-}
-
-pub struct ServiceByteCodeGenerator<'a> {
-    service: &'a Service,
-}
-
-impl<'a> ServiceByteCodeGenerator<'a> {
-    pub fn new(service: &'a Service) -> Self {
-        Self { service }
-    }
-
-    pub fn process_service(&self) -> Result<Vec<Instruction>, ByteCodeError> {
-        let mut code = Vec::new();
-        code.push(Instruction::StoreVar(
-            "name".into(),
-            self.service.name.clone(),
-        ));
-        code.push(Instruction::Jump("main".into()));
-        for method in &self.service.methods {
-            let method_generator = MethodByteCodeGenerator::new(method);
-            let method_code = method_generator.process_method()?;
-            code.extend(method_code);
-        }
-
-        code.push(Instruction::Label("main".into()));
-        for method in &self.service.methods {
-            code.push(Instruction::Jump(format!("{}", method.name)));
-        }
-        code.push(Instruction::Label("end_main".into()));
-        Ok(code)
-    }
-}
-
-pub struct MethodByteCodeGenerator<'a> {
-    method: &'a Method,
-}
-
-impl<'a> MethodByteCodeGenerator<'a> {
-    pub fn new(method: &'a Method) -> Self {
-        Self { method }
-    }
-
-    pub fn process_method(&self) -> Result<Vec<Instruction>, ByteCodeError> {
-        let mut code = Vec::new();
-        code.push(Instruction::Label(format!("{}", self.method.name)));
-
-        if let Some(stdout) = &self.method.stdout {
-            code.push(Instruction::Push(StackValue::String(stdout.clone())));
-            code.push(Instruction::Stdout);
-        }
-
-        if let Some(sleep_ms) = self.method.sleep_ms {
-            code.push(Instruction::Sleep(sleep_ms));
-        }
-
-        if let Some(calls) = &self.method.calls {
-            for call in calls {
-                code.push(Instruction::Push(StackValue::String(call.name.clone())));
-                code.push(Instruction::Push(StackValue::String(call.method.clone())));
-                code.push(Instruction::RemoteCall);
-            }
-        }
-        code.push(Instruction::Jump("main".into()));
-        code.push(Instruction::Label(format!("end_{}", self.method.name)));
-
-        Ok(code)
-    }
-}
-pub struct LogByteCodeGenerator<'a> {
-    task: &'a Task,
-    has_vars: bool,
-}
-
-impl<'a> LogByteCodeGenerator<'a> {
-    pub fn new(task: &'a Task) -> Self {
-        Self {
-            task,
-            has_vars: task.vars.len() > 0,
-        }
-    }
-
-    pub fn process_task(&self) -> Result<Vec<Instruction>, ByteCodeError> {
-        let mut code = Vec::new();
-        code.push(Instruction::StoreVar("name".into(), self.task.name.clone()));
-        code.push(Instruction::StoreVar(
-            "template".into(),
-            self.task.template.clone(),
-        ));
-
-        match &self.task.count {
-            Count::Amount(_) => self.task_with_count(&mut code, self.task)?,
-            Count::Const(val) => {
-                if val == "Infinite" {
-                    self.task_with_infinite_loop(&mut code, self.task)?
-                } else {
-                    return Err(ByteCodeError::UnsupportedConst(val.clone()));
-                }
-            }
-        }
-        Ok(code)
-    }
-
-    fn task_with_infinite_loop(
-        &self,
-        code: &mut Vec<Instruction>,
-        task: &Task,
-    ) -> Result<(), ByteCodeError> {
-        self.generate_var_store_instructions(code, task)?;
-        code.push(Instruction::Label(format!("loop_{}", task.name)));
-        self.generate_print_statement(code, task)?;
-        code.push(Instruction::Jump(format!("loop_{}", task.name)));
-        code.push(Instruction::Label(format!("end_{}", task.name)));
-        Ok(())
-    }
-
-    fn task_with_count(
-        &self,
-        code: &mut Vec<Instruction>,
-        task: &Task,
-    ) -> Result<(), ByteCodeError> {
-        let loop_max_counter = match &task.count {
-            Count::Amount(amount) => amount,
-            Count::Const(val) => {
-                return Err(ByteCodeError::UnsupportedConst(val.clone()));
-            }
-        };
-        self.generate_var_store_instructions(code, task)?;
-        code.push(Instruction::Push(StackValue::Int(*loop_max_counter)));
-        code.push(Instruction::Label(format!("loop_{}", task.name)));
-        code.push(Instruction::Dup);
-        code.push(Instruction::JmpIfZero(format!("end_{}", task.name)));
-        code.push(Instruction::Dec);
-        self.generate_print_statement(code, task)?;
-        code.push(Instruction::Jump(format!("loop_{}", task.name)));
-        code.push(Instruction::Label(format!("end_{}", task.name)));
-        code.push(Instruction::Pop);
-        Ok(())
-    }
-
-    fn generate_var_store_instructions(
-        &self,
-        code: &mut Vec<Instruction>,
-        task: &Task,
-    ) -> Result<(), ByteCodeError> {
-        task.vars.iter().enumerate().for_each(|(index, var)| {
-            code.push(Instruction::StoreVar(format!("var_{}", index), var.clone()));
-        });
-        Ok(())
-    }
-
-    fn generate_print_statement(
-        &self,
-        code: &mut Vec<Instruction>,
-        task: &Task,
-    ) -> Result<(), ByteCodeError> {
-        if self.has_vars {
-            for (index, _var) in task.vars.iter().enumerate() {
-                code.push(Instruction::LoadVar(format!("var_{}", index)));
-                code.push(Instruction::LoadVar("template".into()));
-                code.push(Instruction::Printf);
-
-                match task.severity {
-                    Severity::Info => code.push(Instruction::Stdout),
-                    Severity::Error => code.push(Instruction::Stderr),
-                }
-                code.push(Instruction::Sleep(task.frequency));
-            }
-        } else {
-            code.push(Instruction::LoadVar("template".into()));
-            match task.severity {
-                Severity::Info => code.push(Instruction::Stdout),
-                Severity::Error => code.push(Instruction::Stderr),
-            }
-            code.push(Instruction::Sleep(task.frequency));
-        }
-        Ok(())
-    }
-}
+pub mod error;
+pub mod instruction;
+pub mod log_byte_code;
+pub mod service_byte_code;
 
 #[cfg(test)]
 mod tests {
-    use crate::config::{Call, Config, Method, Service};
-
-    use super::*;
+    use crate::{
+        code_gen::{
+            instruction::{Instruction, StackValue},
+            log_byte_code::LogByteCodeGenerator,
+            service_byte_code::ServiceByteCodeGenerator,
+        },
+        config::{Call, Config, Count, Method, Service, Severity, Task},
+    };
 
     #[test]
     fn test_config_parse() {
